Add PartitionKey to the public API.
