+++ spark/src/main/java/org/apache/iceberg/spark/source/OutputFileFactory.java
@@ -22,6 +22,7 @@ package org.apache.iceberg.spark.source;
 import java.util.UUID;
 import java.util.concurrent.atomic.AtomicInteger;
 import org.apache.iceberg.FileFormat;
+import org.apache.iceberg.PartitionKey;
 import org.apache.iceberg.PartitionSpec;
 import org.apache.iceberg.encryption.EncryptedOutputFile;
 import org.apache.iceberg.encryption.EncryptionManager;
