+++ spark/src/main/java/org/apache/iceberg/spark/source/PartitionedWriter.java
@@ -22,11 +22,13 @@ package org.apache.iceberg.spark.source;
 import java.io.IOException;
 import java.util.Set;
 import org.apache.iceberg.FileFormat;
+import org.apache.iceberg.PartitionKey;
 import org.apache.iceberg.PartitionSpec;
 import org.apache.iceberg.Schema;
 import org.apache.iceberg.io.FileIO;
 import org.apache.iceberg.relocated.com.google.common.collect.Iterables;
 import org.apache.iceberg.relocated.com.google.common.collect.Sets;
+import org.apache.iceberg.spark.SparkSchemaUtil;
 import org.apache.spark.sql.catalyst.InternalRow;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -35,17 +37,19 @@ class PartitionedWriter extends BaseWriter {
   private static final Logger LOG = LoggerFactory.getLogger(PartitionedWriter.class);
 
   private final PartitionKey key;
+  private final InternalRowWrapper wrapper;
   private final Set<PartitionKey> completedPartitions = Sets.newHashSet();
 
   PartitionedWriter(PartitionSpec spec, FileFormat format, SparkAppenderFactory appenderFactory,
                     OutputFileFactory fileFactory, FileIO io, long targetFileSize, Schema writeSchema) {
     super(spec, format, appenderFactory, fileFactory, io, targetFileSize);
     this.key = new PartitionKey(spec, writeSchema);
+    this.wrapper = new InternalRowWrapper(SparkSchemaUtil.convert(writeSchema));
   }
 
   @Override
   public void write(InternalRow row) throws IOException {
-    key.partition(row);
+    key.partition(wrapper.wrap(row));
 
     PartitionKey currentKey = getCurrentKey();
     if (!key.equals(currentKey)) {
